# [WIP] Rewrite marshalling/unmarshalling code, remove io.Writer/Reader support