--- a/bench_xdr_test.go
+++ b/bench_xdr_test.go
@@ -5,9 +5,6 @@
 package xdr_test
 
 import (
-	"bytes"
-	"io"
-
 	"github.com/calmh/xdr"
 )
 
@@ -28,32 +25,26 @@ XDRBenchStruct Structure:
 +-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+
 |                 24 zero bits                  |      I4       |
 +-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+
-|                         Length of Bs0                         |
-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+
 /                                                               /
-\                     Bs0 (variable length)                     \
+\                  Bs0 (length + padded data)                   \
 /                                                               /
 +-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+
-|                         Length of Bs1                         |
-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+
 /                                                               /
-\                     Bs1 (variable length)                     \
+\                  Bs1 (length + padded data)                   \
 /                                                               /
 +-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+
 |                         Number of Is0                         |
 +-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+
-|                              Is0                              |
-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+
-|                         Length of S0                          |
-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+
 /                                                               /
-\                     S0 (variable length)                      \
+|                         Is0 (n items)                         |
 /                                                               /
 +-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+
-|                         Length of S1                          |
+/                                                               /
+\                   S0 (length + padded data)                   \
+/                                                               /
 +-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+
 /                                                               /
-\                     S1 (variable length)                      \
+\                   S1 (length + padded data)                   \
 /                                                               /
 +-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+
 
@@ -72,13 +63,19 @@ struct XDRBenchStruct {
 
 */
 
-func (o XDRBenchStruct) EncodeXDR(w io.Writer) (int, error) {
-	var xw = xdr.NewWriter(w)
-	return o.EncodeXDRInto(xw)
+func (o XDRBenchStruct) XDRSize() int {
+	return 8 + 4 + 4 + 4 +
+		4 + len(o.Bs0) + xdr.Padding(len(o.Bs0)) +
+		4 + len(o.Bs1) + xdr.Padding(len(o.Bs1)) +
+		4 + len(o.Is0)*4 +
+		4 + len(o.S0) + xdr.Padding(len(o.S0)) +
+		4 + len(o.S1) + xdr.Padding(len(o.S1))
 }
 
 func (o XDRBenchStruct) MarshalXDR() ([]byte, error) {
-	return o.AppendXDR(make([]byte, 0, 128))
+	buf := make([]byte, o.XDRSize())
+	m := &xdr.Marshaller{Data: buf}
+	return buf, o.MarshalXDRInto(m)
 }
 
 func (o XDRBenchStruct) MustMarshalXDR() []byte {
@@ -89,128 +86,55 @@ func (o XDRBenchStruct) MustMarshalXDR() []byte {
 	return bs
 }
 
-func (o XDRBenchStruct) AppendXDR(bs []byte) ([]byte, error) {
-	var aw = xdr.AppendWriter(bs)
-	var xw = xdr.NewWriter(&aw)
-	_, err := o.EncodeXDRInto(xw)
-	return []byte(aw), err
-}
-
-func (o XDRBenchStruct) EncodeXDRInto(xw *xdr.Writer) (int, error) {
-	xw.WriteUint64(o.I1)
-	xw.WriteUint32(o.I2)
-	xw.WriteUint16(o.I3)
-	xw.WriteUint8(o.I4)
+func (o XDRBenchStruct) MarshalXDRInto(m *xdr.Marshaller) error {
+	m.MarshalUint64(o.I1)
+	m.MarshalUint32(o.I2)
+	m.MarshalUint16(o.I3)
+	m.MarshalUint8(o.I4)
 	if l := len(o.Bs0); l > 128 {
-		return xw.Tot(), xdr.ElementSizeExceeded("Bs0", l, 128)
+		return xdr.ElementSizeExceeded("Bs0", l, 128)
 	}
-	xw.WriteBytes(o.Bs0)
-	xw.WriteBytes(o.Bs1)
-	xw.WriteUint32(uint32(len(o.Is0)))
+	m.MarshalBytes(o.Bs0)
+	m.MarshalBytes(o.Bs1)
+	m.MarshalUint32(uint32(len(o.Is0)))
 	for i := range o.Is0 {
-		xw.WriteUint32(uint32(o.Is0[i]))
+		m.MarshalUint32(uint32(o.Is0[i]))
 	}
 	if l := len(o.S0); l > 128 {
-		return xw.Tot(), xdr.ElementSizeExceeded("S0", l, 128)
+		return xdr.ElementSizeExceeded("S0", l, 128)
 	}
-	xw.WriteString(o.S0)
-	xw.WriteString(o.S1)
-	return xw.Tot(), xw.Error()
-}
-
-func (o *XDRBenchStruct) DecodeXDR(r io.Reader) error {
-	xr := xdr.NewReader(r)
-	return o.DecodeXDRFrom(xr)
+	m.MarshalString(o.S0)
+	m.MarshalString(o.S1)
+	return m.Error
 }
 
 func (o *XDRBenchStruct) UnmarshalXDR(bs []byte) error {
-	var br = bytes.NewReader(bs)
-	var xr = xdr.NewReader(br)
-	return o.DecodeXDRFrom(xr)
+	u := &xdr.Unmarshaller{Data: bs}
+	return o.UnmarshalXDRFrom(u)
 }
-
-func (o *XDRBenchStruct) DecodeXDRFrom(xr *xdr.Reader) error {
-	o.I1 = xr.ReadUint64()
-	o.I2 = xr.ReadUint32()
-	o.I3 = xr.ReadUint16()
-	o.I4 = xr.ReadUint8()
-	o.Bs0 = xr.ReadBytesMax(128)
-	o.Bs1 = xr.ReadBytes()
-	_Is0Size := int(xr.ReadUint32())
+func (o *XDRBenchStruct) UnmarshalXDRFrom(u *xdr.Unmarshaller) error {
+	o.I1 = u.UnmarshalUint64()
+	o.I2 = u.UnmarshalUint32()
+	o.I3 = u.UnmarshalUint16()
+	o.I4 = u.UnmarshalUint8()
+	o.Bs0 = u.UnmarshalBytesMax(128)
+	o.Bs1 = u.UnmarshalBytes()
+	_Is0Size := int(u.UnmarshalUint32())
 	if _Is0Size < 0 {
 		return xdr.ElementSizeExceeded("Is0", _Is0Size, 0)
+	} else if _Is0Size == 0 {
+		o.Is0 = nil
+	} else {
+		if _Is0Size <= len(o.Is0) {
+			o.Is0 = o.Is0[:_Is0Size]
+		} else {
+			o.Is0 = make([]int32, _Is0Size)
+		}
+		for i := range o.Is0 {
+			o.Is0[i] = int32(u.UnmarshalUint32())
+		}
 	}
-	o.Is0 = make([]int32, _Is0Size)
-	for i := range o.Is0 {
-		o.Is0[i] = int32(xr.ReadUint32())
-	}
-	o.S0 = xr.ReadStringMax(128)
-	o.S1 = xr.ReadString()
-	return xr.Error()
-}
-
-/*
-
-repeatReader Structure:
-
- 0                   1                   2                   3
- 0 1 2 3 4 5 6 7 8 9 0 1 2 3 4 5 6 7 8 9 0 1 2 3 4 5 6 7 8 9 0 1
-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+
-|                        Length of data                         |
-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+
-/                                                               /
-\                    data (variable length)                     \
-/                                                               /
-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+
-
-
-struct repeatReader {
-	opaque data<>;
-}
-
-*/
-
-func (o repeatReader) EncodeXDR(w io.Writer) (int, error) {
-	var xw = xdr.NewWriter(w)
-	return o.EncodeXDRInto(xw)
-}
-
-func (o repeatReader) MarshalXDR() ([]byte, error) {
-	return o.AppendXDR(make([]byte, 0, 128))
-}
-
-func (o repeatReader) MustMarshalXDR() []byte {
-	bs, err := o.MarshalXDR()
-	if err != nil {
-		panic(err)
-	}
-	return bs
-}
-
-func (o repeatReader) AppendXDR(bs []byte) ([]byte, error) {
-	var aw = xdr.AppendWriter(bs)
-	var xw = xdr.NewWriter(&aw)
-	_, err := o.EncodeXDRInto(xw)
-	return []byte(aw), err
-}
-
-func (o repeatReader) EncodeXDRInto(xw *xdr.Writer) (int, error) {
-	xw.WriteBytes(o.data)
-	return xw.Tot(), xw.Error()
-}
-
-func (o *repeatReader) DecodeXDR(r io.Reader) error {
-	xr := xdr.NewReader(r)
-	return o.DecodeXDRFrom(xr)
-}
-
-func (o *repeatReader) UnmarshalXDR(bs []byte) error {
-	var br = bytes.NewReader(bs)
-	var xr = xdr.NewReader(br)
-	return o.DecodeXDRFrom(xr)
-}
-
-func (o *repeatReader) DecodeXDRFrom(xr *xdr.Reader) error {
-	o.data = xr.ReadBytes()
-	return xr.Error()
+	o.S0 = u.UnmarshalStringMax(128)
+	o.S1 = u.UnmarshalString()
+	return u.Error
 }