This is a fairly complete rewrite, to the extent that looking at the diff is fairly meaningless. The new marshalling / unmarshalling code is quite simple though. This is an API change; the tag v1.0.0 retains the old behavior, this one will be tagged as v2.0.0 when it merges.

This makes it so that the only supported operations are unmarshalling a struct from a buffer and marshalling a struct into a buffer. The buffer can be preallocated or one will be allocated, of the right size. The encoded size of a given struct is calculated in advance by the new XDRSize() method.

The advantage is that this skips a shitload of allocations and is much *much* faster.

The unmarshaller uses an ugly trick; any `strings` or `[]byte`s are left as references to the underlying buffer we are unmarshalling from. This means that that buffer cannot be reused while we're holding a reference to the unmarshalled struct - in fact, it should probably be left to go out of scope as soon as possible to avoid mistakes. The advantage is that if we have a struct with five strings in it, that used to be five allocations and five copies. It is now zero allocations and zero copies.

Slices of primitive types like `[]int32` are copied as the underlying representation in memory differs from that in the XDR encoded packet. Existing slices are reused when possible instead of allocating. In case slices are [][]byte or []string the "extra" elements are set to nil/"" to avoid holding on to references to old buffers.

The benchmarks below are for a fairly small struct. Larger ones will show a larger difference.

```
benchmark                    old ns/op     new ns/op     delta
BenchmarkThisMarshal-8       585           251           -57.09%
BenchmarkThisUnmarshal-8     728           110           -84.89%

benchmark                    old allocs     new allocs     delta
BenchmarkThisMarshal-8       4              1              -75.00%
BenchmarkThisUnmarshal-8     7              0              -100.00%

benchmark                    old bytes     new bytes     delta
BenchmarkThisMarshal-8       464           144           -68.97%
BenchmarkThisUnmarshal-8     240           0             -100.00%
```